ADDED: Function to Sort Data As Per Population Store into Json format

+++ src/main/java/com/analyser/dao/CensusDAO.java
@@ -4,7 +4,7 @@ import com.analyser.dto.StateCensusCsv;
 
 public class CensusDAO {
     public String state;
-    public int population;
+    public long population;
     public double areaInSqKm;
     public double densityPerSqKm;
     public String stateCode;
